Remove kernels (for now)
