+++ Cargo.toml
@@ -12,7 +12,6 @@ members = [
     "lib/bindings/c",
     "lib/bindings/python/codegen",
     "lib/engines/*",
-    "lib/kvbm-kernels"
 ]
 # Exclude certain packages that are slow to build and we don't ship as flagship
 # features from default build, but keep them in workspace for convenience.
@@ -48,7 +47,6 @@ dynamo-async-openai = { path = "lib/async-openai", version = "0.6.0", features =
     "rustls",
 ] }
 dynamo-parsers = { path = "lib/parsers", version = "0.6.0" }
-dynamo-kvbm-kernels = { path = "lib/kvbm-kernels" }
 
 # External dependencies
 anyhow = { version = "1" }
