--- lib/kvbm-kernels/Cargo.toml
@@ -1,27 +1,0 @@
-# SPDX-FileCopyrightText: Copyright (c) 2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
-# SPDX-License-Identifier: Apache-2.0
-
-[package]
-name = "dynamo-kvbm-kernels"
-version.workspace = true
-edition.workspace = true
-authors.workspace = true
-license.workspace = true
-repository.workspace = true
-build = "build.rs"
-
-[lib]
-name = "dynamo_kvbm_kernels"
-crate-type = ["rlib", "cdylib"]
-
-[features]
-default = []
-python-bindings = ["pyo3"]
-
-[dependencies]
-cudarc = { workspace = true }
-once_cell = "1.19"
-pyo3 = { version = "0.20", optional = true, features = ["extension-module"] }
-
-[build-dependencies]
-cc = "1.0"
