--- lib/kvbm-kernels/pyproject.toml
@@ -1,26 +1,0 @@
-# SPDX-FileCopyrightText: Copyright (c) 2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
-# SPDX-License-Identifier: Apache-2.0
-
-[build-system]
-requires = ["maturin>=1.5,<2"]
-build-backend = "maturin"
-
-[project]
-name = "kvbm_kernels"
-version = "0.1.0"
-description = "CUDA tensor permutation and packing kernels with Python bindings"
-requires-python = ">=3.8"
-readme = "README.md"
-authors = [{ name = "Ryan", email = "ryan@example.com" }]
-license = { text = "Proprietary" }
-classifiers = [
-    "Programming Language :: Rust",
-    "Programming Language :: Python",
-    "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3 :: Only",
-    "Operating System :: POSIX :: Linux",
-]
-
-[tool.maturin]
-features = ["python-bindings"]
-module-name = "dynamo_kvbm_kernels"
