--- lib/kvbm-kernels/src/lib.rs
@@ -1,12 +1,0 @@
-// SPDX-FileCopyrightText: Copyright (c) 2024-2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-pub mod tensor_kernels;
-
-pub use tensor_kernels::{
-    BlockLayout, OperationalCopyBackend, OperationalCopyDirection, TensorDataType,
-    block_from_universal, operational_copy, universal_from_block,
-};
-
-#[cfg(feature = "python-bindings")]
-mod python;
