+++ lib/llm/Cargo.toml
@@ -21,7 +21,7 @@ testing-full = ["testing-cuda", "testing-nixl"]
 testing-cuda = ["dep:cudarc"]
 testing-nixl = ["dep:nixl-sys"]
 testing-etcd = []
-block-manager = ["dep:nixl-sys", "dep:cudarc", "dep:ndarray", "dep:nix", "dep:aligned-vec", "dep:dynamo-kvbm-kernels"]
+block-manager = ["dep:nixl-sys", "dep:cudarc", "dep:ndarray", "dep:nix", "dep:aligned-vec"]
 cuda = ["dep:cudarc"]
 integration = ["dynamo-runtime/integration"]
 
@@ -99,8 +99,6 @@ nixl-sys = { version = "=0.6.0", optional = true }
 cudarc = { workspace = true, optional = true }
 ndarray = { version = "0.16", optional = true }
 nix = { version = "0.26", optional = true }
-# repo
-dynamo-kvbm-kernels = { workspace = true, optional = true }
 
 
 # protocols
