+++ lib/llm/src/block_manager/v2.rs
@@ -1,5 +1,6 @@
 // SPDX-FileCopyrightText: Copyright (c) 2024-2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
+pub mod kernels;
 pub mod memory;
 pub mod physical;
