+++ lib/llm/src/block_manager/v2/physical/transfer/executor/cuda.rs
@@ -5,12 +5,10 @@
 
 use super::TransferContext;
 use super::{PhysicalLayout, TransferStrategy};
+use crate::block_manager::v2::kernels::OperationalCopyBackend;
 use crate::block_manager::v2::physical::transfer::context::TransferCompleteNotification;
 use anyhow::{Result, anyhow};
 use cudarc::driver::result as cuda_result;
-use dynamo_kvbm_kernels::tensor_kernels::{
-    OperationalCopyBackend, OperationalCopyDirection, TensorDataType,
-};
 use std::ops::Range;
 
 // #[cfg(test)]
@@ -304,163 +302,17 @@ fn execute_d2d(
     Ok(())
 }
 
-/// Check if layouts are compatible for kernel-based operational copy.
-///
-/// Compatible when at least one side is fully contiguous (Operational).
-/// The kernel works with any combination of FC and LayerWise layouts.
-fn are_layouts_kernel_compatible(src: &PhysicalLayout, dst: &PhysicalLayout) -> bool {
-    let src_fc = src.layout().is_fully_contiguous();
-    let dst_fc = dst.layout().is_fully_contiguous();
-
-    // At least one must be FC (operational) - the other can be either FC or LW
-    src_fc || dst_fc
-}
-
-/// Map dtype width in bytes to TensorDataType enum.
-fn map_dtype(dtype_width_bytes: usize) -> Option<TensorDataType> {
-    match dtype_width_bytes {
-        2 => Some(TensorDataType::F16), // Default to F16 for 2-byte types
-        4 => Some(TensorDataType::F32),
-        8 => Some(TensorDataType::F64),
-        _ => None,
-    }
-}
-
-/// Execute device-to-device transfer using operational copy kernels.
-///
-/// This uses the kvbm-kernels operational_copy for FC↔LW or FC↔FC transfers.
-/// Falls back to memcpy-based path on incompatibility or errors.
+/// TODO: For now, we've stubbed this out just so we can merge.
+/// For now, we'll always just fall back to memcpy.
 #[cfg_attr(test, allow(dead_code))]
 pub(crate) fn try_execute_operational_kernel(
-    src: &PhysicalLayout,
-    dst: &PhysicalLayout,
-    src_block_ids: &[usize],
-    dst_block_ids: &[usize],
-    layers: Range<usize>,
-    stream: &cudarc::driver::CudaStream,
-    backend: OperationalCopyBackend,
+    _src: &PhysicalLayout,
+    _dst: &PhysicalLayout,
+    _src_block_ids: &[usize],
+    _dst_block_ids: &[usize],
+    _layers: Range<usize>,
+    _stream: &cudarc::driver::CudaStream,
+    _backend: OperationalCopyBackend,
 ) -> Result<()> {
-    // Check compatibility
-    if !are_layouts_kernel_compatible(src, dst) {
-        return Err(anyhow!("Layouts not compatible for kernel-based copy"));
-    }
-
-    let src_layout = src.layout();
-    let _dst_layout = dst.layout();
-
-    // Map dtype
-    let dtype = map_dtype(src_layout.dtype_width_bytes()).ok_or_else(|| {
-        anyhow!(
-            "Unsupported dtype width: {}",
-            src_layout.dtype_width_bytes()
-        )
-    })?;
-
-    // Determine which side is operational (FC)
-    let src_is_operational = src_layout.is_fully_contiguous();
-    let direction = if src_is_operational {
-        OperationalCopyDirection::OperationalToBlock
-    } else {
-        OperationalCopyDirection::BlockToOperational
-    };
-
-    let (operational_layout, block_layout) = if src_is_operational {
-        (src, dst)
-    } else {
-        (dst, src)
-    };
-
-    // Compute kernel parameters
-    let nl = layers.len();
-    let no = src_layout.outer_dim();
-    let inner = src_layout.page_size() * src_layout.inner_dim();
-    let elem_size = src_layout.dtype_width_bytes();
-    let num_blocks = src_block_ids.len();
-
-    // Build pointer tables
-    // For block layout: flat array of pointers [block][layer][outer]
-    // For operational layout: array of base pointers per block
-    let mut block_ptrs_host: Vec<usize> = Vec::with_capacity(num_blocks * nl * no);
-    let mut operational_ptrs_host: Vec<usize> = Vec::with_capacity(num_blocks);
-
-    let (block_src_ids, block_dst_ids) = if src_is_operational {
-        (dst_block_ids, src_block_ids)
-    } else {
-        (src_block_ids, dst_block_ids)
-    };
-
-    for (&block_src_id, &block_dst_id) in block_src_ids.iter().zip(block_dst_ids.iter()) {
-        // Collect block pointers for this block across all layers and outer dims
-        for layer_id in layers.clone() {
-            for outer_id in 0..no {
-                let region = block_layout.memory_region(block_src_id, layer_id, outer_id)?;
-                block_ptrs_host.push(region.addr());
-            }
-        }
-
-        // Collect operational base pointer for this block
-        // For FC layout, compute base for the first layer in range
-        let op_base = operational_layout.memory_region(block_dst_id, layers.start, 0)?;
-        operational_ptrs_host.push(op_base.addr());
-    }
-
-    // Allocate device memory for pointer tables
-    let block_ptrs_device_raw =
-        unsafe { cuda_result::malloc_sync(block_ptrs_host.len() * std::mem::size_of::<usize>())? };
-    let op_ptrs_device_raw = unsafe {
-        cuda_result::malloc_sync(operational_ptrs_host.len() * std::mem::size_of::<usize>())?
-    };
-
-    // Copy pointer tables to device
-    unsafe {
-        let block_ptrs_slice = std::slice::from_raw_parts(
-            block_ptrs_host.as_ptr() as *const u8,
-            block_ptrs_host.len() * std::mem::size_of::<usize>(),
-        );
-        cuda_result::memcpy_htod_async(
-            block_ptrs_device_raw,
-            block_ptrs_slice,
-            stream.cu_stream(),
-        )?;
-
-        let op_ptrs_slice = std::slice::from_raw_parts(
-            operational_ptrs_host.as_ptr() as *const u8,
-            operational_ptrs_host.len() * std::mem::size_of::<usize>(),
-        );
-        cuda_result::memcpy_htod_async(op_ptrs_device_raw, op_ptrs_slice, stream.cu_stream())?;
-    }
-
-    // Launch kernel
-    let status = unsafe {
-        dynamo_kvbm_kernels::tensor_kernels::operational_copy(
-            block_ptrs_host.as_ptr() as *const *const std::ffi::c_void,
-            block_ptrs_device_raw as usize as *const *const std::ffi::c_void,
-            operational_ptrs_host.as_ptr() as *const *mut std::ffi::c_void,
-            op_ptrs_device_raw as usize as *const *const std::ffi::c_void,
-            num_blocks,
-            nl,
-            no,
-            inner,
-            elem_size,
-            dtype,
-            direction,
-            backend,
-            stream.cu_stream() as cudarc::runtime::sys::cudaStream_t,
-        )
-    };
-
-    // Free device allocations
-    unsafe {
-        let _ = cuda_result::free_async(block_ptrs_device_raw, stream.cu_stream());
-        let _ = cuda_result::free_async(op_ptrs_device_raw, stream.cu_stream());
-    }
-
-    if status != cudarc::runtime::sys::cudaError::cudaSuccess {
-        return Err(anyhow!(
-            "Kernel launch failed with CUDA error: {:?}",
-            status
-        ));
-    }
-
-    Ok(())
+    anyhow::bail!("Not implemented.");
 }
